Rename dsl package to core
